requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy[asyncio]>=2.0.0
aiomysql>=0.2.0
redis>=5.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
openai>=1.3.0
PyYAML>=6.0
python-dotenv>=1.0.0
pymysql>=1.1.0
pdfplumber>=0.10.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1
rank-bm25>=0.2.2
faiss-cpu>=1.7.4
jieba>=0.42.1